now shows start message at the beginning bugfixes: dot on minutes field; no route being displayed; last vs next item

+++ app/src/main/java/flaskoski/rs/smartmuseum/activity/FeaturePreferencesActivity.kt
@@ -69,7 +69,7 @@ class FeaturePreferencesActivity : AppCompatActivity(), FeaturesListAdapter.OnSh
         if(!ApplicationProperties.userNotDefinedYet()){
             txt_username.setText(ApplicationProperties.user?.name)
             txt_hh.setText((ApplicationProperties.user!!.timeAvailable/60.0).toInt().toString())
-            txt_mm.setText((ApplicationProperties.user!!.timeAvailable%60).toString())
+            txt_mm.setText((ApplicationProperties.user!!.timeAvailable%60).toInt().toString())
         }
 
     }
